Create BaseApiController as start of splitting monster ApiController.  Create UserApiController.
