+++ src/main/java/gov/cdc/izgateway/xform/ApiController.java
@@ -2,8 +2,8 @@ package gov.cdc.izgateway.xform;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import gov.cdc.izgateway.security.AccessControlRegistry;
+import gov.cdc.izgateway.xform.api.BaseApiController;
 import gov.cdc.izgateway.xform.model.*;
 import gov.cdc.izgateway.xform.model.Mapping;
 import gov.cdc.izgateway.xform.security.Roles;
@@ -14,20 +14,17 @@ import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.*;
 import java.util.logging.Level;
-import java.util.stream.Collectors;
 
 @Log
 @RestController
 @Lazy(false)
-public class ApiController {
+public class ApiController extends BaseApiController {
     private final OrganizationService organizationService;
     private final PipelineService pipelineService;
     private final SolutionService solutionService;
@@ -36,7 +33,6 @@ public class ApiController {
     private final OperationService operationService;
     private final OperationPreconditionFieldService operationPreconditionFieldService;
     private final GroupRoleMappingService groupRoleMappingService;
-    private final UserService userService;
     private final AccessControlService accessControlService;
 
     @Value("${xform.allow-delete-via-api}")
@@ -53,7 +49,6 @@ public class ApiController {
             OperationPreconditionFieldService operationPreconditionFieldService,
             GroupRoleMappingService groupRoleMappingService,
             AccessControlService accessControlService,
-            UserService userService,
             AccessControlRegistry registry
     ) {
         this.organizationService = organizationService;
@@ -65,7 +60,6 @@ public class ApiController {
         this.operationPreconditionFieldService = operationPreconditionFieldService;
         this.groupRoleMappingService = groupRoleMappingService;
         this.accessControlService = accessControlService;
-        this.userService = userService;
         registry.register(this);
     }
 
@@ -219,22 +213,6 @@ public class ApiController {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
-    
-    @RolesAllowed({Roles.ADMIN})
-    @GetMapping("/api/v1/users")
-    public ResponseEntity<String> getUsersList(
-            @RequestParam(required = false) String nextCursor,
-            @RequestParam(required = false) String prevCursor,
-            @RequestParam(defaultValue = "false") Boolean includeInactive,
-            @RequestParam(defaultValue = "10") int limit
-    ) {
-        try {
-            return processList(userService.getList(), nextCursor, prevCursor, includeInactive, limit);
-        } catch (JsonProcessingException e) {
-            log.log(Level.SEVERE, e.getMessage(), e);
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
 
     @RolesAllowed({Roles.ADMIN})
     @GetMapping("/api/v1/group-role-mappings/{uuid}")
@@ -319,49 +297,6 @@ public class ApiController {
         }
 
         accessControlService.delete(uuid);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-    
-    @RolesAllowed({Roles.ADMIN})
-    @GetMapping("/api/v1/users/{uuid}")
-    public ResponseEntity<User> getUserByUUID(@PathVariable UUID uuid) {
-        User entity = userService.getObject(uuid);
-        if (entity == null) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<>(entity, HttpStatus.OK);
-    }
-
-    @RolesAllowed({Roles.ADMIN})
-    @PutMapping("/api/v1/users/{uuid}")
-    public ResponseEntity<User> updateUser(
-            @PathVariable UUID uuid,
-            @RequestBody @Valid User updatedUser
-    ) {
-        updatedUser.setId(uuid);
-        userService.update(updatedUser);
-        return new ResponseEntity<>(updatedUser, HttpStatus.OK);
-    }
-
-    @RolesAllowed({Roles.ADMIN})
-    @PostMapping("/api/v1/users")
-    public ResponseEntity<User> createUser(
-            @Valid @RequestBody() User user
-    ) {
-        userService.create(user);
-        return new ResponseEntity<>(user, HttpStatus.OK);
-    }
-
-    @RolesAllowed({Roles.ADMIN})
-    @DeleteMapping("/api/v1/users/{uuid}")
-    public ResponseEntity<User> deleteUser(
-            @PathVariable UUID uuid
-    ) {
-        if (Boolean.FALSE.equals(allowDelete)) {
-            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
-        }
-
-        userService.delete(uuid);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
@@ -573,52 +508,6 @@ public class ApiController {
 
         operationPreconditionFieldService.delete(uuid);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-
-    private <T extends BaseModel> ResponseEntity<String> processList(List<T> allEntityList, String nextCursor, String prevCursor, Boolean includeInactive, int limit) throws JsonProcessingException {
-        ObjectMapper mapper = new ObjectMapper();
-        Map<String, Object> returnMap = new HashMap<>();
-        int min = 0;
-        int max = limit;
-        String hasMore = "true";
-
-        List<T> filteredEntityList = filterList(includeInactive, allEntityList);
-
-        for (int i = 0; i < filteredEntityList.size(); i++) {
-            T newEntity = filteredEntityList.get(i);
-            if (newEntity.getId().toString().equals(nextCursor)) {
-                min = i + 1;
-                max = i + limit + 1;
-            } else if (newEntity.getId().toString().equals(prevCursor)) {
-                min = i - limit - 1;
-                max = i;
-            }
-        }
-        if (max > filteredEntityList.size()) {
-            max = filteredEntityList.size();
-            hasMore = "false";
-        }
-
-        if (min < 0) {
-            min = 0;
-            hasMore = "false";
-        }
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        returnMap.put("data", filteredEntityList.subList(min, max));
-        returnMap.put("has_more", hasMore);
-        return new ResponseEntity<>(mapper.writeValueAsString(returnMap), headers, HttpStatus.OK);
-    }
-
-    private <T extends BaseModel> List<T> filterList(Boolean includeInactive, List<T> allList) {
-        if (Boolean.FALSE.equals(includeInactive)) {
-            return allList.stream()
-                    .filter(T::getActive)
-                    .collect(Collectors.toList());
-        } else {
-            return new ArrayList<>(allList);
-        }
     }
 
 }
